style: format with biome

+++ Justfile
@@ -1,5 +1,9 @@
 set quiet := true
 
+personal_settings := "true"
+
+#───────────────────────────────────────────────────────────────────────────────
+
 # INFO uses personal vaults saved in `~/.config/perma-repos.csv`
 build_and_push_to_local_vaults:
     #!/usr/bin/env zsh
@@ -11,12 +15,17 @@ build_and_push_to_local_vaults:
         > /tmp/theme.css
 
     # PUSH TO LOCAL VAULTS
-    cut -d, -f2 "$HOME/.config/perma-repos.csv" |
-        sed "s|^~|$HOME|" |
-        xargs -I {} find {} -maxdepth 1 -name ".obsidian" | # only vaults
-        xargs dirname |                                     # vault root
-        xargs -I {} cp -f "/tmp/theme.css" "{}/.obsidian/themes/Shimmering Focus/theme.css"
-    rm -f /tmp/theme.css
+    if [[ "{{ personal_settings }}" == "true" ]]; then
+        cut -d, -f2 "$HOME/.config/perma-repos.csv" |
+            sed "s|^~|$HOME|" |
+            xargs -I {} find {} -maxdepth 1 -name ".obsidian" | # only vaults
+            xargs dirname |                                     # vault root
+            xargs -I {} cp -f "/tmp/theme.css" "{}/.obsidian/themes/Shimmering Focus/theme.css"
+        rm -f /tmp/theme.css
+        open -a "Obsidian" # macOS specific opener
+    fi
 
-    # OPEN OBSIDIAN
-    open -a "Obsidian" # macOS specific opener
+format_and_check:
+    #!/usr/bin/env zsh
+    [[ "{{ personal_settings }}" != "true" ]] && return 0
+    $HOME/.local/share/nvim/mason/bin/biome check --write --error-on-warnings --log-kind="compact"
